[S] (+) create and join ws utils fn
